+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from './exemplify/services/source.service';
-export * from './exemplify/directives/add-example.directive';
+export * from './exemplify/directives/exemplify.directive';
 export * from './exemplify/exemplify.module';